Configure docsrs

+++ src/half.rs
@@ -1,4 +1,6 @@
 use crate::{impl_shader_layout_array, impl_shader_layout_raw};
+#[cfg(feature = "half")]
+use half::f16;
 
-impl_shader_layout_raw!(half::f16);
-impl_shader_layout_array!(half::f16);
+impl_shader_layout_raw!(f16);
+impl_shader_layout_array!(f16);
